pyproject.toml
[project]
name = "titanic"
description = "titanic example"
version = "0.1.0"
authors = [{ name = "Data Minded", email = "contact@dataminded.be" }]
requires-python = ">=3.9"
dependencies = [
    "awswrangler",
    "pandas<2",
    "scikit-learn",
    "mlflow",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "jupyterlab",
    "pytest",
    "seaborn<0.12",
]

[build-system]
requires = ["setuptools>=62.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
